requirements.txt
certifi==2018.1.18
chardet==3.0.4
click==6.7
Django==2.0.3
Flask==0.12.2
Flask-Uploads==0.2.1
idna==2.6
image==1.5.20
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
Pillow==5.0.0
pymongo==3.6.1
pytz==2018.3
requests==2.18.4
urllib3==1.22
Werkzeug==0.14.1
yagmail==0.10.212
gunicorn

